fix(datetime): centralize datetime conversions across all forms
Create shared datetime utility functions and fix all datetime handling
across the application to properly convert between backend UTC and
user's local timezone.

## Problem

Multiple components were handling datetime conversions inconsistently:
- EventWizard: Had local conversion functions
- EventTokenModal: Directly assigned UTC to datetime-local input
- OrganizationTokenModal: Directly assigned UTC to datetime-local input
- DateTimePicker: Used naive string slicing
- TierForm: Had duplicate conversion functions

All suffered from the same bug: not properly converting UTC to local time.

## Solution

1. Created `src/lib/utils/datetime.ts` with shared utilities:
   - `toDateTimeLocal()`: Converts backend UTC to user's local timezone
   - `toISOString()`: Converts datetime-local to ISO with timezone offset

2. Updated all components to use shared utilities:
   - EventWizard: Replaced inline functions with imports
   - EventTokenModal: Added proper UTC→local conversion
   - OrganizationTokenModal: Added proper UTC→local conversion
   - DateTimePicker: Replaced naive slicing with proper conversion

## Benefits

- **Single source of truth**: All datetime conversions in one place
- **Consistent behavior**: Same logic across entire app
- **Easier maintenance**: Fix once, applies everywhere
- **Type safety**: Shared TypeScript definitions
- **Proper timezone handling**: UTC from backend, local in browser

## Affected Fields

All datetime fields now properly handle timezones:
- Event: start, end, rsvp_before, check_in_starts_at, check_in_ends_at
- Tokens: expires_at
- Tiers: sales_start_at, sales_end_at
- Future: member_since, staff_since, created_at, updated_at
