+++ src/lib/components/tokens/OrganizationTokenModal.svelte
@@ -21,6 +21,7 @@
 	import { RadioGroup, RadioGroupItem } from '$lib/components/ui/radio-group';
 	import { AlertCircle, Loader2 } from 'lucide-svelte';
 	import { durationOptions } from '$lib/utils/tokens';
+	import { toDateTimeLocal, toISOString } from '$lib/utils/datetime';
 
 	interface Props {
 		open: boolean;
@@ -55,7 +56,7 @@
 				grantsMembership = token.grants_membership ?? false;
 				grantsStaffStatus = token.grants_staff_status ?? false;
 				membershipTierId = token.membership_tier || '';
-				expiresAt = token.expires_at || '';
+				expiresAt = toDateTimeLocal(token.expires_at);
 				duration = '1440'; // Not used in edit mode
 			} else {
 				// Reset to defaults for create
@@ -77,7 +78,7 @@
 			const updateData: OrganizationTokenUpdateSchema = {
 				name: name || null,
 				max_uses: maxUses,
-				expires_at: expiresAt || null,
+				expires_at: toISOString(expiresAt),
 				grants_membership: grantsMembership,
 				grants_staff_status: grantsStaffStatus,
 				membership_tier_id: grantsMembership && membershipTierId ? membershipTierId : null
